refactor(test): rename to mock
Also disable the no-mock rule because it seems to be badly supported for
ESM.
