+++ __tests__/file.js
@@ -1,6 +1,8 @@
+/* eslint-disable jest/no-mocks-import */
+
 import { jest } from "@jest/globals";
 
-import { github } from "../__fixtures__/@actions/github.js";
+import { github } from "../__mocks__/@actions/github.js";
 jest.unstable_mockModule("@actions/github", () => github);
 
 const { File, ParseError } = await import("../src/file.js");
