+++ __tests__/link.js
@@ -1,6 +1,8 @@
+/* eslint-disable jest/no-mocks-import */
+
 import { jest } from "@jest/globals";
 
-import { github } from "../__fixtures__/@actions/github.js";
+import { github } from "../__mocks__/@actions/github.js";
 jest.unstable_mockModule("@actions/github", () => github);
 
 const crypto = { hash: jest.fn() };
